requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-braces==1.15.0
django-filter==23.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
ecdsa==0.18.0
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.3
oauthlib==3.2.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wrapt==1.15.0
